api/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-talisman==1.1.0
werkzeug==3.0.3
pandas==2.2.2
openpyxl==3.1.4
anthropic==0.40.0
